refactor: items of layout schema
Utilizing the if-then-else for items provides more meaningful error messages when validating JSON. Previously, if one property was missing, you would see that error for all layout item types.
